pyproject.toml
[tool.poetry]
name = "gazprom-test"
version = "0.1.0"
description = ""
authors = ["surendoy <wladkorostilev@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.1"
alembic = "^1.13.1"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
bcrypt = "^4.1.3"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.30" }
asyncpg = "^0.29.0"
gunicorn = "^22.0.0"
apscheduler = "^3.10.4"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
uvicorn = "^0.30.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
